backend/requirements.txt
# TenderCortex - Dependencies
# Python 3.11+

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.27.0

# Configuration & Validation
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
langchain-groq>=0.1.0
langchain-huggingface>=0.0.1

# Vector Database (Qdrant in-memory - zero maintenance)
qdrant-client>=1.9.0
langchain-qdrant>=0.1.0

# Document Processing
pypdf>=4.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
dateparser>=1.2.0

# Data Analysis & Visualization (QuanT)
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
networkx>=3.2.0
